% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_distance.R
\name{angular_distance}
\alias{angular_distance}
\title{angular_distance}
\usage{
angular_distance(
  input_dir,
  allout_dir,
  subject_name,
  sel_normal_sample,
  all_normal_samples,
  new_sample_names,
  sample_name_change,
  bootstrap,
  bscut,
  bsreps
)
}
\description{
Generate angular distance matrices, heatmaps, and trees. Optionally generate bootstrapped data.
}
