% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{angular_distance_heatmap}
\alias{angular_distance_heatmap}
\title{angular_distance_heatmap}
\usage{
angular_distance_heatmap(
  filepath_old,
  filepath_new,
  table_filepath_old,
  table_filepath_new,
  subject_name,
  Z,
  markerset,
  usedsamps,
  sel_normal_sample,
  new_rownames,
  sample_name_change,
  markergroup,
  replicate
)
}
\description{
Create heatmap for angular distrance w.r.t the normal
}
