% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clustermap.R
\name{create_clustermap}
\alias{create_clustermap}
\title{create_clustermap}
\usage{
create_clustermap(
  output_dir,
  subject_name,
  dm_df,
  max_replicate_th,
  sample_names
)
}
\description{
create_clustermap
}
