% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{create_distance_matrix}
\alias{create_distance_matrix}
\title{create_distance_matrix}
\usage{
create_distance_matrix(
  sample_names,
  final_marker_names,
  repre_replicates,
  raw_dist
)
}
\description{
Create JSM distance matrix for representative replicate. Only markers with no missing samples are used. (Deprecated?)
}
