% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{create_JSM_matrix}
\alias{create_JSM_matrix}
\title{create_JSM_matrix}
\usage{
create_JSM_matrix(norm_df)
}
\description{
Calculate Jensen-Shannon distance for each pair of samples of the given marker and creates a matrix with the pairwise distances norm_df: normalized probability distributions of the input polyg data. Returns: Jensen-Shannon distance matrix
}
\author{
Alexander Gorelick <agorelick@mgh.harvard.edu>
}
