% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_marker_files.R
\name{create_marker_files}
\alias{create_marker_files}
\title{create_marker_files}
\usage{
create_marker_files(
  sample_table_file,
  peak_table_file,
  markers_file,
  mixes_file,
  sizing_artifacts_file = NA,
  pdir = ".",
  hcut = 0.015,
  thres = 20,
  intensityfilter = 0.1
)
}
\description{
create_marker_files
}
