% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_phylogenetic_trees_with_bootstrap_values.R
\name{create_phylogenetic_trees_with_bootstrap_values}
\alias{create_phylogenetic_trees_with_bootstrap_values}
\title{create_phylogenetic_trees_with_bootstrap_values}
\usage{
create_phylogenetic_trees_with_bootstrap_values(
  subject_name,
  sel_normal_sample,
  all_normal_samples,
  pdir,
  overwrite,
  sample_exclusion_th,
  max_replicate_th,
  intfraction_cutoff = 0.1,
  abovebelow_cutoff = 0.15,
  distcut = 0.06,
  file_filter = 0.1,
  rob_th = 0.05,
  sample_name_change = T,
  colorfile_header = F,
  KLM = F,
  bscut = 50
)
}
\description{
create_phylogenetic_trees_with_bootstrap_values
}
