% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_distribution_heatmap.R
\name{distance_distribution_heatmap}
\alias{distance_distribution_heatmap}
\title{distance_distribution_heatmap}
\usage{
distance_distribution_heatmap(
  filepath_old,
  filepath_new,
  subject_name,
  repre_replicates,
  track_lengths,
  final_marker_names,
  markers,
  max_replicate_th,
  sel_normal_sample,
  all_normal_samples,
  sample_names,
  new_rownames,
  add_info_dir,
  rob_th
)
}
\description{
Perform average/UPGMA linkage on the JSM distance matrix and create distrance distribution heatmap. use final_marker_names and sample_names with possible exclusion of certain samples and markers
}
