% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{generate_angular_distance_trees}
\alias{generate_angular_distance_trees}
\title{generate_angular_distance_trees}
\usage{
generate_angular_distance_trees(
  angular_dist_w_root,
  sel_normal_sample,
  colorfile,
  sample_name_change,
  ad_tree_w_root_dir,
  markerset,
  usedsamps,
  ns,
  power,
  new_sample_names,
  meanlen_diff_normal,
  bootstrap,
  bscut,
  subject_name,
  markergroup,
  replicate,
  tree_type = "NJ"
)
}
\description{
build a tree with the angular distance
}
