% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_output_filepath}
\alias{get_output_filepath}
\title{get_output_filepath}
\usage{
get_output_filepath(subject_name, title, suffix, output_dir, max_replicate_th)
}
\description{
Generate filename and path
}
