% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/phylogenetic_tree_with_boostrap_sampleNameChange.R
\name{phylogenetic_tree_with_boostrap_sampleNameChange}
\alias{phylogenetic_tree_with_boostrap_sampleNameChange}
\title{phylogenetic_tree_with_boostrap_sampleNameChange}
\usage{
phylogenetic_tree_with_boostrap_sampleNameChange(
  subject_name,
  pdir,
  output_dir,
  sample_exclusion_th,
  max_replicate_th,
  sel_normal_sample,
  all_normal_samples,
  colorfile_header,
  KLM,
  allout_dir,
  bscut,
  method
)
}
\description{
phylogenetic_tree_with_boostrap_sampleNameChange
}
