\name{polyG-package}
\alias{polyG-package}
\alias{polyG}
\docType{package}
\title{
\packageTitle{polyG}
}
\description{
\packageDescription{polyG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{polyG}
\packageIndices{polyG}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{polyG}
Maintainer: \packageMaintainer{polyG}
}

