% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_rep_replicates_create_dist_matrix.R
\name{sel_rep_replicates_create_dist_matrix}
\alias{sel_rep_replicates_create_dist_matrix}
\title{sel_rep_replicates_create_dist_matrix}
\usage{
sel_rep_replicates_create_dist_matrix(
  subject_name,
  sel_normal_sample,
  all_normal_samples,
  pdir = ".",
  overwrite = F,
  sample_exclusion_th = 0.15,
  max_replicate_th = 0.11,
  intfraction_cutoff = 0.1,
  abovebelow_cutoff = 0.15,
  distcut = 0.06,
  file_filter = 0.1,
  rob_th = 0.05,
  sample_name_change = T
)
}
\description{
sel_rep_replicates_create_dist_matrix
}
