% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_representative_replicates.R
\name{select_representative_replicates}
\alias{select_representative_replicates}
\title{select_representative_replicates}
\usage{
select_representative_replicates(
  marker,
  raw_dist,
  intensity,
  sample_replicates,
  repre_replicates,
  max_replicate_th,
  all_sample_names,
  outlist
)
}
\description{
select_representative_replicates
}
