% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\title{subject}
\usage{
subject(
  subject_name,
  pdir,
  output_dir,
  sample_exclusion_th,
  max_replicate_th,
  sel_normal_sample,
  all_normal_samples
)
}
\description{
subject
}
