% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyG.R
\name{polyG}
\alias{polyG}
\title{polyG}
\usage{
polyG(
  input_dir,
  results_dir,
  max_replicate_th,
  sample_name_change,
  bootstrap,
  bscut,
  seed = NA
)
}
\description{
Main script to process polyG data, from raw data input to angular distance data.
}
