% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{select_representative_replicates}
\alias{select_representative_replicates}
\title{select_representative_replicates}
\usage{
select_representative_replicates(
  marker,
  raw_dist,
  intensity,
  sample_replicates,
  repre_replicates,
  max_replicate_th,
  all_sample_names,
  outlist
)
}
\description{
Select most representative replicates. First check if all of them agree, then select the pair which minimizes the provided distance and then select the replicate that maximizes the intensity
}
