% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\title{subject}
\usage{
subject(
  input_dir,
  allout_dir,
  subject_name,
  data_path,
  sample_exclusion_th,
  sel_normal_sample,
  all_normal_samples,
  new_sample_names
)
}
\description{
Process the raw poly-G marker data for this subject. Results will be used to in angular_distance() function.
}
